feat(settings): a pipeline can be retired from the screen, and put back
Retiring a pipeline and restoring one have worked at the API since #5155
and the screen offered neither, so the only way to stop offering a
pipeline nobody uses was to call the endpoint by hand. Worse, an admin
who did that put it somewhere the UI could not show them again: the card
read GET /pipelines with no include_archived, so a retired pipeline
simply vanished from the one screen that administers them.

The card now asks for archived rows on a request OF ITS OWN, and that is
the half a wider query would have got wrong: the deal board, the
stage-automation picker and the lead qualifier all read the same
["pipelines","all"] cache entry, so widening it would have put every
retired pipeline back in the three places retiring one removes it from.

Retiring is pipeline:DELETE, the verb stage removal already takes, and
restoring is pipeline:update — two grants, because the server checks two.
The default pipeline's control is drawn and refused with the server's own
remedy beside it rather than hidden: the reason is the information, and
it names something the reader can do from that same row.

The pipeline UI leaves settings.tsx on the way. It was over the length
cap under a frozen waiver, and the ladder splits from the pipeline the
way the records do: settings.stages.tsx names no pipeline type and
imports nothing from settings.pipelines.tsx, so the direction reads from
the imports. The waiver drops by 499 lines.

archive.tsx's claim that no resource has a restore endpoint is no longer
true and says so. The rule it was justifying survives the correction:
that hook owns a confirm, an invalidation and a toast saying a record is
gone, and a restore is none of those.
