+++ frontend/src/screens/settings.tsx
@@ -89,9 +89,7 @@ import { CompanyContextCard } from "./company-context";
 import { ConnectedAgentsCard } from "./connected-agents";
 import { ConnectorsCard } from "./connectors";
 import { ConsumerMailDomainsCard } from "./consumer-mail-domains";
-import { CreateAction, type CreateField, CreateRecordModal } from "./create";
 import { CustomFieldsAdmin } from "./customfields";
-import { EditAction } from "./edit";
 import { EmbedReindexCard } from "./embedreindex";
 import { EntityRef } from "./entityref";
 import { ExtensionAccessCard } from "./extension-access";
@@ -126,7 +124,7 @@ import { FxRatesCard, ModelCostsCard } from "./rates";
 import { RecordRolesCard } from "./recordroles";
 import { RestrictedRecordsCard } from "./restrictedrecords";
 import { RetentionCard } from "./retention";
-import { StageExitCriteria } from "./settings.exitcriteria";
+import { PipelinesCard } from "./settings.pipelines";
 import { StageAutomationCard } from "./settings.stageautomation";
 import { SignInMethodsCard } from "./sign-in-methods";
 import { TagVocabularyCard } from "./tagadmin";
@@ -1869,497 +1867,6 @@ function ResetDataCard() {
           onChange={(event) => setTyped(event.target.value)}
         />
       </ConfirmModal>
-    </Panel>
-  );
-}
-
-type Pipeline = components["schemas"]["Pipeline"];
-type Stage = components["schemas"]["Stage"];
-
-// The 3 shared scalar fields between create and edit pipeline forms.
-function pipelineFields(t: ReturnType<typeof useT>): CreateField[] {
-  return [
-    { key: "name", label: "pipeline.name", required: true },
-    {
-      key: "is_default",
-      label: "pipeline.default",
-      type: "select",
-      required: true,
-      options: [
-        { value: "false", label: t("pipeline.notDefault") },
-        { value: "true", label: t("pipeline.default") },
-      ],
-    },
-    { key: "position", label: "pipeline.position", type: "number" },
-  ];
-}
-
-// Coerces a form value (CreateAction's values are strings; EditAction's
-// update callback widens to Record<string, unknown> so a screen COULD prefill
-// non-string values) down to the trimmed string this form always produces —
-// mirrors deals.tsx's mapDealUpdate `str` helper, keeping both create's and
-// edit's transports on the one map function without an `as` cast.
-function str(v: unknown): string {
-  return typeof v === "string" ? v.trim() : "";
-}
-
-function mapPipelineBody(v: Record<string, unknown>) {
-  return {
-    name: str(v.name),
-    is_default: v.is_default === "true",
-    position: v.position ? Number(str(v.position)) : 0,
-  };
-}
-
-// Narrows the form's free-text semantic value into the Stage enum WITHOUT a
-// cast (mirrors deals.tsx's forecastCategory) — an unrecognized value falls
-// back to "open" rather than shipping a bad literal to the wire.
-function stageSemantic(v: unknown): Stage["semantic"] {
-  switch (v) {
-    case "won":
-      return "won";
-    case "lost":
-      return "lost";
-    default:
-      return "open";
-  }
-}
-
-// UpdateStageRequest carries no pipeline_id (a stage never moves pipelines
-// via this form) while CreateStageRequest requires one — so this returns
-// only the fields the two requests share, and the create transport adds
-// pipeline_id on top.
-function mapStageBody(v: Record<string, unknown>) {
-  return {
-    name: str(v.name),
-    position: v.position ? Number(str(v.position)) : 0,
-    semantic: stageSemantic(v.semantic),
-    win_probability: v.win_probability ? Number(str(v.win_probability)) : 0,
-  };
-}
-
-function stageFields(t: ReturnType<typeof useT>): CreateField[] {
-  return [
-    { key: "name", label: "stage.name", required: true },
-    { key: "position", label: "pipeline.position", type: "number" },
-    {
-      key: "semantic",
-      label: "stage.semantic",
-      type: "select",
-      required: true,
-      options: [
-        { value: "open", label: t("stage.semOpen") },
-        { value: "won", label: t("stage.semWon") },
-        { value: "lost", label: t("stage.semLost") },
-      ],
-    },
-    { key: "win_probability", label: "stage.winProb", type: "number" },
-  ];
-}
-
-// Localized badge for a stage's semantic — open/won/lost each render as a
-// short label rather than the raw enum value.
-function stageSemanticLabel(
-  semantic: Stage["semantic"],
-  t: ReturnType<typeof useT>,
-): string {
-  if (semantic === "won") {
-    return t("stage.semWon");
-  }
-  if (semantic === "lost") {
-    return t("stage.semLost");
-  }
-  return t("stage.semOpen");
-}
-
-// Tone-less Badge shares the card-inset background it sits on (both resolve
-// to var(--bgCard)) — the semantic pill needs an explicit tone to be visible.
-function stageSemanticTone(
-  semantic: Stage["semantic"],
-): "success" | "danger" | "accent" {
-  switch (semantic) {
-    case "won":
-      return "success";
-    case "lost":
-      return "danger";
-    default:
-      return "accent"; // open
-  }
-}
-
-// The bespoke per-pipeline "new stage" trigger: CreateAction's testid
-// (`new-record`) can't disambiguate multiple pipelines on one screen, so
-// this composes the same Button + CreateRecordModal pieces directly rather
-// than adding new form infra.
-function StageCreate({ pipelineId }: Readonly<{ pipelineId: string }>) {
-  const t = useT();
-  const [open, setOpen] = useState(false);
-  const queryClient = useQueryClient();
-  const mutation = useMutation({
-    mutationFn: async (values: Record<string, string>) => {
-      const { data, error } = await api.POST("/stages", {
-        body: { ...mapStageBody(values), pipeline_id: pipelineId },
-      });
-      if (error) {
-        throwProblem(error);
-      }
-      return data;
-    },
-    onSuccess: () => {
-      setOpen(false);
-      queryClient.invalidateQueries({ queryKey: ["pipelines"] });
-    },
-  });
-  return (
-    <>
-      <Button
-        small
-        data-testid={`new-stage-${pipelineId}`}
-        onClick={() => setOpen(true)}
-      >
-        {t("stage.new")}
-      </Button>
-      <CreateRecordModal
-        open={open}
-        onClose={() => setOpen(false)}
-        title={t("stage.new")}
-        fields={stageFields(t)}
-        pending={mutation.isPending}
-        error={mutation.isError ? problemMessageOf(mutation.error, t) : null}
-        onSubmit={(values) => mutation.mutate(values)}
-      />
-    </>
-  );
-}
-
-// The removal half of the bounded stage surface. Both refusals are the
-// server's — a stage still holding deals, and the terminal won/lost pair —
-// so this asks and then shows what it was told rather than pre-judging
-// from the row: the refusal names the deals standing in the way, which is
-// the part an admin acts on, and a board read a minute ago would name the
-// wrong ones.
-function StageRemove({
-  stage,
-  returnFocusTo,
-}: Readonly<{ stage: Stage; returnFocusTo: () => HTMLElement | null }>) {
-  const t = useT();
-  const [open, setOpen] = useState(false);
-  const queryClient = useQueryClient();
-  // Removal is pipeline:delete, not the pipeline:update everything else on
-  // this card runs on — the server gates it that way, so a principal who
-  // may add and rename stages but not remove one is not shown a control
-  // that could only ever answer 403. Read before the early return: the
-  // hooks a render performs must not depend on the answer.
-  const canRemove = useCanWrite("pipeline", "delete");
-  const remove = useMutation({
-    mutationFn: async () => {
-      const { error } = await api.DELETE("/stages/{id}", {
-        params: { path: { id: stage.id } },
-      });
-      if (error) {
-        throwProblem(error);
-      }
-    },
-    onSuccess: async () => {
-      // The refetched pipelines FIRST, then the dialog: closing it hands
-      // focus back to a list that must no longer hold this row.
-      await queryClient.invalidateQueries({ queryKey: ["pipelines"] });
-      setOpen(false);
-    },
-  });
-  // A refusal is about the workspace's state, not the dialog's: reopening
-  // must ask again rather than reprint what the last attempt was told.
-  const close = () => {
-    remove.reset();
-    setOpen(false);
-  };
-  if (!canRemove) {
-    return null;
-  }
-  return (
-    <>
-      {/* Ghost, not danger. The dialog behind it is where the danger lives —
-          its confirm is the red one — and a trigger that shouts as loudly as
-          the act it only ASKS about put six solid red buttons in one pipeline,
-          which is the shout a reader stops reading. */}
-      <Button
-        small
-        data-testid={`remove-stage-${stage.id}`}
-        onClick={() => setOpen(true)}
-      >
-        {t("stage.remove")}
-      </Button>
-      <ConfirmModal
-        open={open}
-        onClose={close}
-        title={t("stage.removeTitle")}
-        confirmLabel={t("stage.removeConfirm")}
-        confirmVariant="danger"
-        pending={remove.isPending}
-        error={remove.isError ? problemMessageOf(remove.error, t) : null}
-        onConfirm={() => remove.mutate()}
-        // The stage list, not the trigger: a successful removal unmounts
-        // the row this button lives in, so there is nothing to hand focus
-        // back to (design-system/confirmmodal).
-        returnFocusTo={returnFocusTo}
-      >
-        <p className="t-caption">
-          {t("stage.removeBody", { name: stage.name })}
-        </p>
-      </ConfirmModal>
-    </>
-  );
-}
-
-function StageRow({
-  stage,
-  canEdit,
-  t,
-  returnFocusTo,
-}: Readonly<{
-  stage: Stage;
-  canEdit: boolean;
-  t: ReturnType<typeof useT>;
-  returnFocusTo: () => HTMLElement | null;
-}>) {
-  const { locale } = useLocale();
-  return (
-    // The four tracks (name, semantic badge, probability, edit) live in
-    // settings.css, where they can have a phone breakpoint. Inline, the three
-    // fixed tracks plus the Edit button left about 60px for a stage name that
-    // could not wrap, and it painted straight over the badge beside it.
-    <li className="stage-row">
-      <span className="stage-name">{stage.name}</span>
-      <Badge tone={stageSemanticTone(stage.semantic)}>
-        {stageSemanticLabel(stage.semantic, t)}
-      </Badge>
-      <span className="t-mono t-caption">
-        {formatNumber(stage.win_probability, locale)}%
-      </span>
-      {/* Each control carries its own verb — editing a stage is
-          pipeline:update, removing one is pipeline:delete — so a role
-          holding one without the other still sees the one it may use. */}
-      <span className="stage-verbs">
-        {canEdit && (
-          <EditAction<Stage>
-            label={t("stage.edit")}
-            savedMessage={(saved) => t("record.saveDone", { name: saved.name })}
-            invalidate="pipelines"
-            recordKey="stage"
-            record={{
-              id: stage.id,
-              name: stage.name,
-              position: String(stage.position),
-              semantic: stage.semantic,
-              win_probability: String(stage.win_probability),
-            }}
-            fields={stageFields(t)}
-            update={async (values) => {
-              const { data, error } = await api.PATCH("/stages/{id}", {
-                params: { path: { id: stage.id } },
-                body: mapStageBody(values),
-              });
-              if (error) {
-                throwProblem(error);
-              }
-              return data;
-            }}
-          />
-        )}
-        <StageRemove stage={stage} returnFocusTo={returnFocusTo} />
-      </span>
-      {/* The criteria sit UNDER the stage's own line rather than beside it:
-          a stage carries three to six of them, each with a label, a key, two
-          badges and two verbs, which is a list and not an answer that fits in
-          a track. */}
-      <div className="stage-criteria">
-        <Disclosure summary={t("stage.criteria.title")}>
-          <StageExitCriteria
-            stageId={stage.id}
-            semantic={stage.semantic}
-            canEdit={canEdit}
-          />
-        </Disclosure>
-      </div>
-    </li>
-  );
-}
-
-// One pipeline as one row: its name on the left, and under it what the pipeline
-// IS — default or not — the verbs that change it, and the stage ladder itself.
-//
-// Stacked, because the ladder IS the subject rather than an answer to a question
-// that fits beside it: three to six stages, each carrying a name, a semantic
-// badge, a probability and two verbs of its own. The pipeline's own name is the
-// row's LABEL, which is what puts it at the same x as every other naming on the
-// page — it used to be an inner heading, drawn one step larger than the card
-// title above it.
-function PipelineRow({
-  pipeline,
-  canEdit,
-  t,
-}: Readonly<{
-  pipeline: Pipeline;
-  canEdit: boolean;
-  t: ReturnType<typeof useT>;
-}>) {
-  const stageList = useRef<HTMLUListElement>(null);
-  const stages = [...(pipeline.stages ?? [])].sort(
-    (a, b) => a.position - b.position,
-  );
-  return (
-    <SettingRow
-      label={pipeline.name}
-      layout="stack"
-      control={
-        <div className="form-stack settingrow-measure">
-          {/* What this pipeline IS, and the verbs that change it, above the
-              ladder they act on. */}
-          <div className="pipeline-standing">
-            <Badge tone={pipeline.is_default ? "success" : undefined}>
-              {pipeline.is_default
-                ? t("pipeline.default")
-                : t("pipeline.notDefault")}
-            </Badge>
-            {canEdit && (
-              <>
-                <EditAction<Pipeline>
-                  label={t("pipeline.edit")}
-                  savedMessage={(saved) =>
-                    t("record.saveDone", { name: saved.name })
-                  }
-                  invalidate="pipelines"
-                  recordKey="pipeline"
-                  record={{
-                    id: pipeline.id,
-                    name: pipeline.name,
-                    is_default: String(pipeline.is_default),
-                    position: String(pipeline.position),
-                  }}
-                  fields={pipelineFields(t)}
-                  update={async (values) => {
-                    const { data, error } = await api.PATCH("/pipelines/{id}", {
-                      params: { path: { id: pipeline.id } },
-                      body: mapPipelineBody(values),
-                    });
-                    if (error) {
-                      throwProblem(error);
-                    }
-                    return data;
-                  }}
-                />
-                <StageCreate pipelineId={pipeline.id} />
-              </>
-            )}
-          </div>
-          {/* tabIndex -1 so a removal can hand focus to the list it changed:
-              the row's own Remove button is gone by then, and focus dropped to
-              <body> leaves a screen-reader user at the top of the document. */}
-          <ul ref={stageList} tabIndex={-1} className="stage-rows">
-            {stages.map((stage) => (
-              <StageRow
-                key={stage.id}
-                stage={stage}
-                canEdit={canEdit}
-                t={t}
-                returnFocusTo={() => stageList.current}
-              />
-            ))}
-          </ul>
-        </div>
-      }
-    />
-  );
-}
-
-// D-8: Settings → Pipelines config. Reads via the SAME ["pipelines","all"]
-// key the deals screen's plural selector uses (an array shape, distinct
-// from DealScreen's single-pipeline ["pipelines"] cache entry) — any
-// mutation here invalidates the ["pipelines"] prefix, so both shapes stay
-// fresh. The list itself is readable by everyone; only the write affordances are
-// gated, and the server stays the RBAC authority. Three of the five seeded roles
-// hold pipeline READ and no write verb at all, so for most readers this card is
-// the read-only case rather than an edge of it — which is why it states that
-// posture once instead of leaving a reader to infer it from absent buttons.
-export function PipelinesCard() {
-  const t = useT();
-  // Adding a pipeline is pipeline:create. Everything else here — renaming a
-  // pipeline, adding a stage, editing one, reordering — is pipeline:update,
-  // including the stage CREATE affordance: a stage is not its own RBAC object,
-  // so adding one is an update to the pipeline that owns it.
-  const canCreate = useCanWrite("pipeline", "create");
-  const canEdit = useCanWrite("pipeline", "update");
-  const query = useQuery({
-    queryKey: ["pipelines", "all"],
-    queryFn: async () => {
-      const { data, error } = await api.GET("/pipelines", {
-        params: { query: {} },
-      });
-      if (error) {
-        throwProblem(error);
-      }
-      return data.data;
-    },
-  });
-  return (
-    <Panel
-      title={t("settings.pipelines")}
-      // Adding a pipeline is four inputs committed together, so the header
-      // keeps the verb and the dialog keeps the form. It sits in the header
-      // band rather than as a trailing row: a row's label would repeat the
-      // button beside it, and a card-level create verb is the header's job
-      // everywhere else on these pages. Absent without the create grant,
-      // exactly as each Edit verb is — the read-only posture is stated once
-      // below.
-      titleAction={
-        canCreate && (
-          <CreateAction
-            label={t("pipeline.new")}
-            invalidate="pipelines"
-            screen="settings"
-            create={async (values) => {
-              const { data, error } = await api.POST("/pipelines", {
-                body: { ...mapPipelineBody(values), stages: [] },
-              });
-              if (error) {
-                throwProblem(error);
-              }
-              return data;
-            }}
-            fields={pipelineFields(t)}
-          />
-        )
-      }
-    >
-      <PanelBody>
-        <p className="settings-panel-sub">{t("settings.pipelinesSub")}</p>
-        {/* Said once, at the top, rather than annotating each absent control —
-            the rule in design-system/README.md. A reader holding one of the two
-            verbs can see for themselves which controls they got. */}
-        {!canCreate && !canEdit && (
-          <p className="settings-panel-sub">
-            {t("settings.pipelinesReadOnly")}
-          </p>
-        )}
-        <SettingList>
-          <QueryGate
-            pendingLabel={t("settings.pipelines")}
-            query={query}
-            empty={(pipelines) => pipelines.length === 0}
-          >
-            {(pipelines) =>
-              pipelines.map((pipeline) => (
-                <PipelineRow
-                  key={pipeline.id}
-                  pipeline={pipeline}
-                  canEdit={canEdit}
-                  t={t}
-                />
-              ))
-            }
-          </QueryGate>
-        </SettingList>
-      </PanelBody>
     </Panel>
   );
 }
