+++ scripts/fe-file-length-waivers.txt
@@ -70,7 +70,7 @@
 2576 frontend/src/screens/companies.tsx
 2645 frontend/src/screens/deals.test.tsx
 2750 frontend/src/screens/compose.tsx
-2811 frontend/src/screens/settings.tsx
+2312 frontend/src/screens/settings.tsx
 2833 frontend/src/screens/leads.test.tsx
 3331 frontend/src/screens/compose.test.tsx
 4497 frontend/src/screens/deals.tsx
